% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Game that returns Fizz for anything divisible by 3, Buzz for anything
and FizzBuzz for anything divisible by 3 and 5.}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A vector of natural numbers}
}
\value{
A Character vector of numbers and/or words
}
\description{
Takes a vector of values and returns a vector of FizzBuzz responses
}
